+++ handlers/orders.js
@@ -232,6 +232,7 @@ handlers.getMySoferOrders = async function (req, res, args) {
         `SELECT o.id, o.client, o.ref, o.loc_incarcare, o.loc_descarcare, o.km,
                 o.data_incarcare, o.data_descarcare,
                 o.firma_incarcare, o.firma_descarcare,
+                o.sosit_incarcare_at, o.incarcat_at, o.sosit_descarcare_at, o.descarcat_at,
                 o.rendszam_camion, o.rendszam_remorca, o.status,
                 o.handover_status, o.handover_type, o.handover_loc,
                 o.finalized_at,